requirements.txt
aiohttp==3.11.18
pymodbus==3.8.1
PyYAML==6.0.2
